test: move reusable stubs to __fixtures__ in src
