+++ packages/babel-plugin/src/__fixtures__/mixins/objects.js
