+++ packages/babel-plugin/src/__fixtures__/mixins/simple.js
