+++ packages/babel-plugin/src/__fixtures__/mixins/strings.js
