+++ packages/babel-plugin/src/__tests__/module-traversal.test.tsx
@@ -37,7 +37,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import colors from './stubs/objects';
+      import colors from '../__fixtures__/mixins/objects';
 
       <div css={{ color: colors.primary }} />
     `
@@ -51,7 +51,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import color from './stubs/simple';
+      import color from '../__fixtures__/mixins/simple';
 
       <div css={{ color }} />
     `
@@ -65,7 +65,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { primary } from './stubs/simple';
+      import { primary } from '../__fixtures__/mixins/simple';
 
       <div css={{ color: primary }} />
     `
@@ -79,7 +79,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { colors } from './stubs/objects';
+      import { colors } from '../__fixtures__/mixins/objects';
 
       <div css={{ color: colors.primary }} />
     `
@@ -107,7 +107,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { style } from './stubs/objects';
+      import { style } from '../__fixtures__/mixins/objects';
 
       <div css={style} />
     `
@@ -121,7 +121,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { style } from './stubs/objects';
+      import { style } from '../__fixtures__/mixins/objects';
 
       <div css={{ color: 'blue', ...style }} />
     `
@@ -135,7 +135,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { styleInlining } from './stubs/objects';
+      import { styleInlining } from '../__fixtures__/mixins/objects';
 
       <div css={styleInlining} />
     `
@@ -149,7 +149,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { styleInlining } from './stubs/objects';
+      import { styleInlining } from '../__fixtures__/mixins/objects';
 
       <div css={{ ...styleInlining }} />
     `
@@ -163,7 +163,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { styleModuleInlining } from './stubs/objects';
+      import { styleModuleInlining } from '../__fixtures__/mixins/objects';
 
       <div css={{ ...styleModuleInlining }} />
     `
@@ -177,7 +177,7 @@ describe('module traversal', () => {
       `
       import '@compiled/core';
       import React from 'react';
-      import { styleModuleInlining } from './stubs/objects';
+      import { styleModuleInlining } from '../__fixtures__/mixins/objects';
 
       <div css={styleModuleInlining} />
     `
@@ -190,7 +190,7 @@ describe('module traversal', () => {
     const result = transform(
       `
       import '@compiled/core';
-      import { reexport } from './stubs/reexport';
+      import { reexport } from '../__fixtures__/mixins/reexport';
 
       <div css={{ color: reexport }} />
     `
@@ -203,7 +203,7 @@ describe('module traversal', () => {
     const result = transform(
       `
       import '@compiled/core';
-      import { objectReexport } from './stubs/reexport';
+      import { objectReexport } from '../__fixtures__/mixins/reexport';
 
       <div css={{ color: objectReexport.foo }} />
     `
@@ -216,7 +216,7 @@ describe('module traversal', () => {
     const result = transform(
       `
       import '@compiled/core';
-      import { bold } from './stubs/strings';
+      import { bold } from '../__fixtures__/mixins/strings';
 
       <div css={bold} />
     `
@@ -229,7 +229,7 @@ describe('module traversal', () => {
     const result = transform(
       `
       import '@compiled/core';
-      import { italics } from './stubs/strings';
+      import { italics } from '../__fixtures__/mixins/strings';
 
       <div css={[italics]} />
     `
@@ -242,7 +242,7 @@ describe('module traversal', () => {
     const result = transform(
       `
       import '@compiled/core';
-      import { danger } from './stubs/strings';
+      import { danger } from '../__fixtures__/mixins/strings';
 
       <div css={[danger]} />
     `
@@ -256,7 +256,7 @@ describe('module traversal', () => {
       transform(
         `
         import '@compiled/core';
-        import { cantStaticallyEvaluate } from './stubs/objects';
+        import { cantStaticallyEvaluate } from '../__fixtures__/mixins/objects';
 
         <div css={[cantStaticallyEvaluate]} />
       `
@@ -269,7 +269,7 @@ describe('module traversal', () => {
       transform(
         `
       import '@compiled/core';
-      import { cantStaticallyEvaluate } from './stubs/objects';
+      import { cantStaticallyEvaluate } from '../__fixtures__/mixins/objects';
 
       <div css={{ ...cantStaticallyEvaluate }} />
     `
@@ -281,7 +281,7 @@ describe('module traversal', () => {
     const result = transform(
       `
       import '@compiled/core';
-      import { colorMixin } from './stubs/objects';
+      import { colorMixin } from '../__fixtures__/mixins/objects';
 
       <div css={{ ':hover': colorMixin() }} />
     `
@@ -294,7 +294,7 @@ describe('module traversal', () => {
     const result = transform(
       `
       import '@compiled/core';
-      import { colorMixin } from './stubs/objects';
+      import { colorMixin } from '../__fixtures__/mixins/objects';
 
       const colors = colorMixin();
